requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.49.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyJWT==2.10.1
pycparser==3.0
python-decouple==3.8
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
